requirements.txt
python-telegram-bot
python-dotenv
praw
unidecode